+++ comicbox/formats/comic_info/schema.py
@@ -151,7 +151,7 @@ class ComicInfoSubSchema(XmlSubHeadSchema):
     Letterer = XmlStringSetField(as_string=True)
     CoverArtist = XmlStringSetField(as_string=True)
     Editor = XmlStringSetField(as_string=True)
-    Translator = XmlStringField()
+    Translator = XmlStringSetField(as_string=True)
     # End Role Tags
     Publisher = XmlStringField()
     Imprint = XmlStringField()
@@ -175,9 +175,9 @@ class ComicInfoSubSchema(XmlSubHeadSchema):
         "Page", ListField(Nested(XmlPageInfoSchema), sort_keys=(IMAGE_ATTRIBUTE,))
     )
     CommunityRating = XmlDecimalField()
-    MainCharacterOrTeam = XmlStringSetField(as_string=True)
+    MainCharacterOrTeam = XmlStringField()
     Review = XmlStringField()
-    GTIN = XmlStringSetField(as_string=True)
+    GTIN = XmlStringField()
 
     class Meta(XmlSubHeadSchema.Meta):
         """Schema options."""